momoxrss_python/requirements.txt
# API & serveur
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Base de données
pymongo==4.6.1
motor==3.3.2

# Scheduler
apscheduler==3.10.4

# HTTP & parsing
requests==2.32.3
feedparser==6.0.11
tldextract==5.1.2

# Discord bot
discord.py==2.4.0
PyNaCl==1.6.0

# Validation & modèles
pydantic==2.9.2
pydantic-settings==2.5.2

# Utilitaires
python-dotenv==1.0.1
psutil==5.9.8
